refactor(nats): rename VoiceWorkerRegistry to WorkerRegistry
The registry holds no voice-specific state and is now imported by image
(and soon memory/llm) clients. The voice-prefixed name was misleading —
readers had to inspect the module to confirm there was no hidden coupling.

Closes #805
