+++ src/lyra/nats/nats_tts_client.py
@@ -1,6 +1,6 @@
 """NatsTtsClient — hub-side NATS request-reply client for TTS.
 
-Maintains a ``VoiceWorkerRegistry`` populated from heartbeats, and routes each
+Maintains a ``WorkerRegistry`` populated from heartbeats, and routes each
 synthesis to the least-loaded worker via its per-worker subject
 ``lyra.voice.tts.request.{worker_id}``. Falls back once to the queue-group
 subject (``lyra.voice.tts.request``) if the targeted worker times out.
@@ -18,7 +18,7 @@ from uuid import uuid4
 from nats.aio.client import Client as NATS
 from pydantic import ValidationError
 
-from lyra.nats.voice_health import VoiceWorkerRegistry
+from lyra.nats.worker_registry import WorkerRegistry
 from lyra.tts import SynthesisResult, TtsUnavailableError
 from roxabi_contracts.envelope import CONTRACT_VERSION
 from roxabi_contracts.voice import (
@@ -42,7 +42,7 @@ class NatsTtsClient:
         self._nc = nc
         self._timeout = timeout
         self._cb = NatsCircuitBreaker()
-        self._registry = VoiceWorkerRegistry()
+        self._registry = WorkerRegistry()
         self._hb_sub = None  # set by start
 
     async def start(self) -> None:
